backend/requirements.txt
Django==5.1.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
google-auth==2.48.0
psycopg2-binary==2.9.10
PyJWT==2.11.0
sqlparse==0.5.3
gunicorn==23.0.0
dj-database-url==2.3.0
python-dotenv==1.0.1
whitenoise==6.9.0
Pillow==11.1.0
locust==2.32.6

